split checks into cfchecks, datachecks and missing modules
