+++ tests/test_checks.py
@@ -8,7 +8,12 @@ import pytest
 import xarray as xr
 
 from xclim import set_options
-from xclim.core import checks
+from xclim.core.cfchecks import check_valid
+from xclim.core.cfchecks import check_valid_temperature
+from xclim.core.missing import at_least_n_valid
+from xclim.core.missing import missing_any
+from xclim.core.missing import missing_pct
+from xclim.core.missing import missing_wmo
 from xclim.core.utils import ValidationError
 from xclim.indicators.atmos import tg_mean
 
@@ -25,7 +30,7 @@ TestObj = namedtuple("TestObj", ["test"])
 )
 def test_check_valid_ok(value, expected):
     d = TestObj(value)
-    checks.check_valid(d, "test", expected)
+    check_valid(d, "test", expected)
 
 
 @pytest.mark.parametrize(
@@ -34,7 +39,7 @@ def test_check_valid_ok(value, expected):
 def test_check_valid_raise(value, expected):
     d = TestObj(value)
     with pytest.raises(ValidationError):
-        checks.check_valid(d, "test", expected)
+        check_valid(d, "test", expected)
 
 
 class TestDateHandling:
@@ -83,7 +88,7 @@ def test_cf_compliance_options(tas_series, caplog):
 
     with caplog.at_level(logging.INFO):
         with set_options(cf_compliance="log"):
-            checks.check_valid_temperature(tas, "degK")
+            check_valid_temperature(tas, "degK")
 
             assert all([rec.levelname == "INFO" for rec in caplog.records])
             assert (
@@ -93,13 +98,13 @@ def test_cf_compliance_options(tas_series, caplog):
 
     with pytest.warns(UserWarning, match="Variable has a non-conforming"):
         with set_options(cf_compliance="warn"):
-            checks.check_valid_temperature(tas, "degK")
+            check_valid_temperature(tas, "degK")
 
     with pytest.raises(
         ValidationError, match="Variable has a non-conforming standard_name"
     ):
         with set_options(cf_compliance="raise"):
-            checks.check_valid_temperature(tas, "degK")
+            check_valid_temperature(tas, "degK")
 
 
 class TestMissingAnyFills:
@@ -107,7 +112,7 @@ class TestMissingAnyFills:
         a = np.arange(360.0)
         a[5:10] = np.nan
         ts = tas_series(a)
-        out = checks.missing_any(ts, freq="MS")
+        out = missing_any(ts, freq="MS")
         assert out[0]
         assert not out[1]
 
@@ -115,93 +120,93 @@ class TestMissingAnyFills:
         n = 66
         times = pd.date_range("2001-12-30", freq="1D", periods=n)
         da = xr.DataArray(np.arange(n), [("time", times)])
-        miss = checks.missing_any(da, "MS")
+        miss = missing_any(da, "MS")
         np.testing.assert_array_equal(miss, [True, False, False, True])
 
     def test_missing_years(self):
         n = 378
         times = pd.date_range("2001-12-31", freq="1D", periods=n)
         da = xr.DataArray(np.arange(n), [("time", times)])
-        miss = checks.missing_any(da, "YS")
+        miss = missing_any(da, "YS")
         np.testing.assert_array_equal(miss, [True, False, True])
 
     def test_missing_season(self):
         n = 378
         times = pd.date_range("2001-12-31", freq="1D", periods=n)
         da = xr.DataArray(np.arange(n), [("time", times)])
-        miss = checks.missing_any(da, "Q-NOV")
+        miss = missing_any(da, "Q-NOV")
         np.testing.assert_array_equal(miss, [True, False, False, False, True])
 
     def test_to_period_start(self, tasmin_series):
         a = np.zeros(365) + K2C + 5.0
         a[2] -= 20
         ts = tasmin_series(a)
-        miss = checks.missing_any(ts, freq="AS-JUL")
+        miss = missing_any(ts, freq="AS-JUL")
         np.testing.assert_equal(miss, [False])
 
     def test_to_period_end(self, tasmin_series):
         a = np.zeros(365) + K2C + 5.0
         a[2] -= 20
         ts = tasmin_series(a)
-        miss = checks.missing_any(ts, freq="A-JUN")
+        miss = missing_any(ts, freq="A-JUN")
         np.testing.assert_equal(miss, [False])
 
     def test_month(self, tasmin_series):
         ts = tasmin_series(np.zeros(36))
-        miss = checks.missing_any(ts, freq="YS", month=7)
+        miss = missing_any(ts, freq="YS", month=7)
         np.testing.assert_equal(miss, [False])
 
-        miss = checks.missing_any(ts, freq="YS", month=8)
+        miss = missing_any(ts, freq="YS", month=8)
         np.testing.assert_equal(miss, [True])
 
         with pytest.raises(ValueError, match=r"No data for selected period."):
-            checks.missing_any(ts, freq="YS", month=1)
+            missing_any(ts, freq="YS", month=1)
 
-        miss = checks.missing_any(ts, freq="YS", month=[7, 8])
+        miss = missing_any(ts, freq="YS", month=[7, 8])
         np.testing.assert_equal(miss, [True])
 
         ts = tasmin_series(np.zeros(76))
-        miss = checks.missing_any(ts, freq="YS", month=[7, 8])
+        miss = missing_any(ts, freq="YS", month=[7, 8])
         np.testing.assert_equal(miss, [False])
 
     def test_season(self, tasmin_series):
         ts = tasmin_series(np.zeros(360))
-        miss = checks.missing_any(ts, freq="YS", season="MAM")
+        miss = missing_any(ts, freq="YS", season="MAM")
         np.testing.assert_equal(miss, [False])
 
-        miss = checks.missing_any(ts, freq="YS", season="JJA")
+        miss = missing_any(ts, freq="YS", season="JJA")
         np.testing.assert_array_equal(miss, [True, True])
 
-        miss = checks.missing_any(ts, freq="YS", season="SON")
+        miss = missing_any(ts, freq="YS", season="SON")
         np.testing.assert_equal(miss, [False])
 
     def test_no_freq(self, tasmin_series):
         ts = tasmin_series(np.zeros(360))
 
-        miss = checks.missing_any(ts, freq=None)
+        miss = missing_any(ts, freq=None)
         np.testing.assert_array_equal(miss, False)
 
         t = list(range(31))
         t.pop(5)
         ts2 = ts.isel(time=t)
-        miss = checks.missing_any(ts2, freq=None)
+        miss = missing_any(ts2, freq=None)
         np.testing.assert_array_equal(miss, True)
 
         # With indexer
-        miss = checks.missing_any(ts, freq=None, month=[7])
+        miss = missing_any(ts, freq=None, month=[7])
         np.testing.assert_array_equal(miss, False)
 
-        miss = checks.missing_any(ts2, freq=None, month=[7])
+        miss = missing_any(ts2, freq=None, month=[7])
         np.testing.assert_array_equal(miss, True)
 
     def test_hydro(self):
         fn = Path(TESTS_DATA, "Raven", "q_sim.nc")
         ds = xr.open_dataset(fn)
-        miss = checks.missing_any(ds.q_sim, freq="YS")
+        miss = missing_any(ds.q_sim, freq="YS")
         np.testing.assert_array_equal(miss[:-1], False)
         np.testing.assert_array_equal(miss[-1], True)
 
-        miss = checks.missing_any(ds.q_sim, freq="YS", season="JJA")
+        miss = missing_any(ds.q_sim, freq="YS", season="JJA")
         np.testing.assert_array_equal(miss, False)
 
 
@@ -212,7 +217,7 @@ class TestMissingWMO:
         a[40:45] = np.nan  # Too many consecutive missing values
         a[70:92:2] = np.nan  # Too many non-consecutive missing values
         ts = tas_series(a)
-        out = checks.missing_wmo(ts, freq="MS")
+        out = missing_wmo(ts, freq="MS")
         assert not out[0]
         assert out[1]
         assert out[2]
@@ -221,7 +226,7 @@ class TestMissingWMO:
         a = np.arange(360.0)
         a[5:16] = np.nan  # Number of missing values under acceptable limit in a month
         ts = tas_series(a)
-        out = checks.missing_wmo(ts, freq="QS-JAN")
+        out = missing_wmo(ts, freq="QS-JAN")
         np.testing.assert_array_equal(out, [True, False, False, True])
 
 
@@ -231,7 +236,7 @@ class TestMissingPct:
         a[5:7] = np.nan  # Number of missing values under acceptable limit in a month
         a[40:45] = np.nan  # Too many missing values
         ts = tas_series(a)
-        out = checks.missing_pct(ts, freq="MS", tolerance=0.1)
+        out = missing_pct(ts, freq="MS", tolerance=0.1)
         assert not out[0]
         assert out[1]
 
@@ -242,5 +247,5 @@ class TestAtLeastNValid:
         a[5:10] = np.nan  # Number of missing values under acceptable limit in a month
         a[40:55] = np.nan  # Too many missing values
         ts = tas_series(a)
-        out = checks.at_least_n_valid(ts, freq="MS", n=20)
+        out = at_least_n_valid(ts, freq="MS", n=20)
         np.testing.assert_array_equal(out[:2], [False, True])
