+++ tests/test_options.py
@@ -4,7 +4,7 @@
 import pytest
 
 from xclim import set_options
-from xclim.core.checks import MissingBase
+from xclim.core.missing import MissingBase
 from xclim.core.options import OPTIONS
 from xclim.core.options import register_missing_method
