+++ xclim/indicators/atmos/_temperature.py
@@ -2,7 +2,7 @@
 import abc
 
 from xclim import indices
-from xclim.core import checks
+from xclim.core.cfchecks import check_valid
 from xclim.core.indicator import Daily
 from xclim.core.indicator import Daily2D
 from xclim.core.units import check_units
@@ -61,33 +61,33 @@ class Tas(Daily):
 
     @staticmethod
     def cfcheck(tas):
-        checks.check_valid(tas, "cell_methods", "*time: mean within days*")
-        checks.check_valid(tas, "standard_name", "air_temperature")
+        check_valid(tas, "cell_methods", "*time: mean within days*")
+        check_valid(tas, "standard_name", "air_temperature")
 
 
 class Tasmin(Tas):
     """Class for univariate indices using min daily temperature as the input."""
 
     def cfcheck(self, tasmin):
-        checks.check_valid(tasmin, "cell_methods", "*time: minimum within days*")
-        checks.check_valid(tasmin, "standard_name", "air_temperature")
+        check_valid(tasmin, "cell_methods", "*time: minimum within days*")
+        check_valid(tasmin, "standard_name", "air_temperature")
 
 
 class Tasmax(Tas):
     """Class for univariate indices using max daily temperature as the input."""
 
     def cfcheck(self, tasmax):
-        checks.check_valid(tasmax, "cell_methods", "*time: maximum within days*")
-        checks.check_valid(tasmax, "standard_name", "air_temperature")
+        check_valid(tasmax, "cell_methods", "*time: maximum within days*")
+        check_valid(tasmax, "standard_name", "air_temperature")
 
 
 class TasminTasmax(Daily2D):
     def cfcheck(self, tasmin, tasmax):
         for da in (tasmin, tasmax):
-            checks.check_valid(da, "standard_name", "air_temperature")
+            check_valid(da, "standard_name", "air_temperature")
 
-        checks.check_valid(tasmin, "cell_methods", "*time: minimum within days*")
-        checks.check_valid(tasmax, "cell_methods", "*time: maximum within days*")
+        check_valid(tasmin, "cell_methods", "*time: minimum within days*")
+        check_valid(tasmax, "cell_methods", "*time: maximum within days*")
         check_units(tasmax, tasmin.attrs["units"])
